+++ src/tree.c
@@ -353,6 +353,9 @@ static const imask_t has_map[T_LAST_TREE_KIND] = {
 
    // T_VIEW_ELEMENT
    (I_IDENT | I_REF | I_SUBKIND | I_TYPE | I_VALUE),
+
+   // T_MATCH_SELECT
+   (I_IDENT | I_VALUE | I_STMTS),
 };
 
 static const char *kind_text_map[T_LAST_TREE_KIND] = {
@@ -392,6 +395,7 @@ static const char *kind_text_map[T_LAST_TREE_KIND] = {
    "T_ALTERNATIVE",     "T_PSL",             "T_VERILOG",
    "T_VIEW_DECL",       "T_PACKAGE_MAP",     "T_COND_EXPR",
    "T_COND_VALUE",      "T_COND_RETURN",     "T_VIEW_ELEMENT",
+   "T_MATCH_SELECT",
 };
 
 static const change_allowed_t change_allowed[] = {